refactor: Recruitment 중복 부분 Components 추가 - RecruitmentList
